+++ rust/src/lib.rs
@@ -49,8 +49,7 @@
 //! and are a good place to start until we have demo apps built in Rust).
 
 pub use mediasoup_types as types;
-#[deprecated(note = "Deprecated in favour of types::data_structures")]
-pub mod data_structures;
+mod data_structures;
 pub(crate) mod fbs;
 mod macros;
 mod messages;
@@ -58,14 +57,9 @@ mod messages;
 pub mod ortc;
 pub mod prelude;
 pub mod router;
-#[deprecated(note = "Deprecated in favour of types::rtp_parameters")]
-pub mod rtp_parameters;
-#[deprecated(note = "Deprecated in favour of types::scalability_modes")]
-pub mod scalability_modes;
-#[deprecated(note = "Deprecated in favour of types::sctp_parameters")]
-pub mod sctp_parameters;
-#[deprecated(note = "Deprecated in favour of types::srtp_parameters")]
-pub mod srtp_parameters;
+mod rtp_parameters;
+mod sctp_parameters;
+mod srtp_parameters;
 pub mod supported_rtp_capabilities;
 pub mod webrtc_server;
 pub mod worker;