+++ rust/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Make `parameters` and `rtcp_feedback` optional in `RtpCodecParameters` and `RtpCodecCapability` during deserialization (PR #1589).
 - Make codec `mime_type` case insensitive during deserialization (PR #1590).
+- Only expose `data_structures`, `rtp_parameters`, `sctp_parameters` and `srtp_parameters` through the `mediasoup-types` crate (PR #1591).
 
 # 0.19.1